Move language getters/setters to rdf_wrapper so that FairWorkflow objects also have them
